requirements.txt
# Job Genie Requirements
python>=3.8
pdfplumber>=0.7.0
python-docx>=0.8.11
fpdf2>=2.5.0
beautifulsoup4>=4.10.0
requests>=2.27.1
PyPDF2>=2.10.0
spacy>=3.2.0
nltk>=3.7
tkinter>=8.6
pillow>=9.0.0